+++ CHANGELOG.md
@@ -1,4 +1,21 @@
 
+## 0.5
+
+### 0.5.0
+
+- [BREAKING CHANGE] Meteor 1.2 is required to work
+- [BREAKING CHANGE] **All absolute paths to modules must start with `/`**
+- [BREAKING CHANGE] `!vars` was replaced by `!exports`
+- [BREAKING CHANGE] Remove backward compatibility for deprecated package syntax `author:package`
+
+- Change in internal module naming (Potentially breaking change)
+- Make use of new build plugins API
+- Update `SystemJS` to 0.19.0
+- Use Meteor Promise polyfill instead of this shipped with SystemJS
+- Improve error handling (stack traces are now easier to read when errors are thrown inside module)
+- Update `babel-compiler` to x
+- Update `babel-runtime` to x
+
 ## 0.4
 
 ### 0.4.1